requirements.txt
streamlit
cohere
python-dotenv
PyPDF2
pandas
altair
